DistilBERT-CRF/requirements.txt
accelerate>=0.34.0
datasets>=2.14.0
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.1.0
pyyaml>=6.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
seaborn>=0.13.0
seqeval>=1.2.2
torch>=2.5.0
torchcrf>=1.1.0
tqdm>=4.66.0
transformers>=4.46.0
umap-learn>=0.5.5
